feat: jwt auth

+++ src/dashboard-service.ts
@@ -3,13 +3,16 @@ import {Content, Dashboard, PrismaClient } from "@prisma/client";
 export class DashboardService{
     constructor(private readonly prisma: PrismaClient){}
 
-    async moveDashboard(dashboardId: string, position: number): Promise<boolean>{
+    async moveDashboard(userId: string, dashboardId: string, position: number): Promise<boolean>{
             const dashboards = await this.prisma.dashboard.findMany({
+                where:{
+                    userId: userId,
+                },
                 orderBy: {
                     position: "asc",
                 },
             });
-            //i can't use a position greater than my array's len
+            //i can't use a position greater than my array's lenn
             if(position >= dashboards.length){
                 return false
             }
@@ -24,7 +27,11 @@ export class DashboardService{
             return true;
         }
     //implementing CRUD api    
-    async createContent(dashboardId: string, text: string){
+    async createContent(userId: string,dashboardId: string, text: string){
+        const dashboard = await this.getDash(userId, dashboardId);
+        if(!dashboard){
+            return;
+        }
         const countContent = await this.prisma.content.count({
             where: {
                 dashboardId: dashboardId,
@@ -39,17 +46,22 @@ export class DashboardService{
         })
     }
 
-    async createDashboard(name: string){
+    async createDashboard(userId: string,name: string){
         const countDashboards = await this.prisma.dashboard.count({});
         return await this.prisma.dashboard.create({
             data: {
                 position: countDashboards,
                 name: name,
+                userId: userId,
             }
         })
     }
 
-    async deleteDashboard(dashboardId: string){
+    async deleteDashboard(userId: string,dashboardId: string){
+        const dashboard = await this.getDash(userId, dashboardId);
+        if(!dashboard){
+            return null;
+        }
         const contentsInDashboard = await this.prisma.content.count({
             where:{
                 dashboardId: dashboardId
@@ -65,7 +77,11 @@ export class DashboardService{
         })
     }
 
-    async deleteContent(dashboardId: string, contentId: string){
+    async deleteContent(userId: string,dashboardId: string, contentId: string){
+        const dashboard = await this.getDash(userId, dashboardId);
+        if(!dashboard){
+            return;
+        }
         return await this.prisma.content.delete({
             where: {
                 id_dashboardId: {
@@ -77,12 +93,26 @@ export class DashboardService{
     }
 
     async moveContent(
+            userId: string,
             contentId: string, 
             position: number, 
             fromDashboardId: string, 
             toDashboardId: string
             ): Promise<boolean>{
-            const fromToSameDashboard = fromDashboardId == toDashboardId;
+            const fromToSameDashboard = fromDashboardId === toDashboardId;
+
+            const fromDashboard = await this.getDash(userId, fromDashboardId);
+            if(!fromDashboard){
+                return false;
+            }
+
+            if(!fromToSameDashboard){
+                const toDashboard = await this.getDash(userId, toDashboardId);
+                if(!toDashboard){
+                    return false;
+                    }
+            }
+
                 //list of contents from start
             const fromContents = await this.prisma.content.findMany({
                 orderBy: {
@@ -127,8 +157,11 @@ export class DashboardService{
             return true;
         }
 
-    getDashboard(){
+    getDashboard(userId: string){
         return this.prisma.dashboard.findMany({
+            where: {
+                userId: userId,
+            },
             orderBy: {
                 position: "asc",
             },
@@ -142,7 +175,7 @@ export class DashboardService{
         });
     }
 
-    async reoderDashboard(dashboards: Dashboard[]){
+    private async reoderDashboard(dashboards: Dashboard[]){
         //prepate all my updates transaction
         const updates = dashboards.map((dashboard, position)=>{
             return this.prisma.dashboard.update({
@@ -158,7 +191,7 @@ export class DashboardService{
         await this.prisma.$transaction(updates)
     }
 
-    async reoderContent(contents: Content[], dashboardId: string){
+    private async reoderContent(contents: Content[], dashboardId: string){
         const updates = contents.map((content, position)=>{
             return this.prisma.content.update({
                 where: {
@@ -171,6 +204,17 @@ export class DashboardService{
             });
         });
         await this.prisma.$transaction(updates)
+    }
+    
+    getDash(userId: string, dashboardId: string){
+        return this.prisma.dashboard.findUnique({
+            where:{
+                id_userId: {
+                    id: dashboardId,
+                    userId: userId,
+                },
+            },
+        });
     }
     
 }
